Move exception-related `util_tiledb` methods to `exception`.
This is the rest of what is in `util_tiledb.py`.
